refactor(components): improve component structure and organization
