+++ main/src/app/(forum)/(categories)/c/[categoryID]/page.tsx
@@ -2,8 +2,8 @@ import { Metadata } from "next";
 import { ChevronRight } from "lucide-react";
 import { ChevronDown } from "lucide-react";
 import { Plus } from "lucide-react";
-import Aside from "@/components/side_info/main";
-import Post from "@/components/Category_Section/post";
+import Aside from "@/components/Forum/side_info/main";
+import Post from "@/components/Forum/Category_Section/post";
 import { v4 as uuidv4 } from "uuid";
 import Link from "next/link";
