+++ main/src/components/Forum/Messages/chat_view.tsx
