+++ main/src/components/Forum/Post_Section/post.tsx
