+++ main/src/components/Forum/main_section/main.tsx
